Revert "Updated installation, organised and formatted the things"
